requirements.txt
pyyaml>=6.0
pandas>=2.0
streamlit>=1.32
pytest>=8.0
python-dateutil>=2.9
